core/ocr_handler.py
#!/usr/bin/env python3
# -*- coding: utf-8 -*-
"""
OCR驗證碼處理模組
支援pytesseract和easyocr兩種OCR引擎
"""

import cv2
import numpy as np
from PIL import Image
import pytesseract
import easyocr
from typing import Optional, List, Tuple
import io
import base64


class OCRHandler:
    """OCR驗證碼處理器"""
    
    def __init__(self, ocr_engine: str = "tesseract"):
        """
        初始化OCR處理器
        支援 'tesseract' 或 'easyocr' 引擎
        """
        pass
    
    def preprocess_image(self, image_data: bytes) -> np.ndarray:
        """
        圖片預處理
        1. 轉換為灰度圖
        2. 去噪處理
        3. 二值化
        4. 形態學操作
        """
        pass
    
    def enhance_image(self, image: np.ndarray) -> np.ndarray:
        """
        增強圖片品質
        1. 調整對比度和亮度
        2. 銳化處理
        3. 去除干擾線
        """
        pass
    
    def extract_text_tesseract(self, image: np.ndarray) -> str:
        """
        使用Tesseract OCR識別文字
        設定OCR參數並返回識別結果
        """
        pass
    
    def extract_text_easyocr(self, image: np.ndarray) -> str:
        """
        使用EasyOCR識別文字
        支援中英文混合識別
        """
        pass
    
    def process_captcha_from_element(self, driver, captcha_element) -> str:
        """
        從網頁元素中處理驗證碼
        1. 截取驗證碼圖片
        2. 圖片預處理
        3. OCR識別
        4. 結果後處理
        """
        pass
    
    def process_captcha_from_url(self, image_url: str) -> str:
        """
        從圖片URL處理驗證碼
        下載圖片並進行OCR識別
        """
        pass
    
    def process_captcha_from_base64(self, base64_data: str) -> str:
        """
        從Base64數據處理驗證碼
        解碼並進行OCR識別
        """
        pass
    
    def validate_result(self, text: str) -> str:
        """
        驗證和清理OCR結果
        1. 移除特殊字符
        2. 長度驗證
        3. 格式檢查（數字/字母/混合）
        """
        pass
    
    def get_multiple_results(self, image: np.ndarray, count: int = 3) -> List[str]:
        """
        獲取多個識別結果
        使用不同參數多次識別，提高準確率
        """
        pass
    
    def save_debug_image(self, image: np.ndarray, filename: str):
        """
        保存除錯圖片
        用於分析OCR處理效果
        """
        pass 