requirements.txt
undetected-chromedriver>=3.5.0
selenium>=4.15.0
pytesseract>=0.3.10
easyocr>=1.7.0
Pillow>=10.0.0
requests>=2.31.0
configparser>=5.3.0
loguru>=0.7.0
webdriver-manager>=4.0.0
tkinter>=8.6.0 